+++ inc/entry.js
@@ -101,6 +101,6 @@
  * This is a callback for any errors and results from running a command.
  *
  * @callback App~commandRunCallback
- * @param {Object|undefined} err - The error (if any) returned from this command
- * @param {String|undefined} res - The response (if any) returned from this command
+ * @param {Object|undefined} [err] - The error (if any) returned from this command
+ * @param {String|undefined} [res] - The response (if any) returned from this command
  */